# public/swagger/v1/swagger.json
{
	"components": {
...
		"schemas": {
...
			"BoundingBox": {
				"properties": {
					"minLatitude": {
						"type": "number",
						"format": "double"
					},
					"maxLatitude": {
						"type": "number",
						"format": "double"
					},
					"minLongitude": {
						"type": "number",
						"format": "double"
					},
					"maxLongitude": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"minLatitude",
					"maxLatitude",
					"minLongitude",
					"maxLongitude"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Type": {
				"properties": {
					"identifier": {
						"type": "string"
					},
					"label": {
						"type": "string"
					}
				},
				"required": [
					"identifier",
					"label"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ComparisonGroup": {
				"properties": {
					"identifier": {
						"type": "string"
					},
					"isCustom": {
						"type": "boolean"
					},
					"label": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/Type"
					}
				},
				"required": [
					"identifier",
					"isCustom",
					"label",
					"type"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IComparisonGroupsGetResponse": {
				"properties": {
					"comparisonGroup-array": {
						"items": {
							"$ref": "#/components/schemas/ComparisonGroup"
						},
						"type": "array"
					}
				},
				"required": [
					"comparisonGroup-array"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Member": {
				"properties": {
					"identifier": {
						"type": "string"
					},
					"label": {
						"type": "string"
					},
					"shortLabel": {
						"type": "string"
					},
					"longLabel": {
						"type": "string"
					},
					"areaType": {
						"$ref": "#/components/schemas/Member"
					},
					"isGroup": {
						"type": "boolean"
					}
				},
				"required": [
					"identifier",
					"label",
					"shortLabel",
					"longLabel"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Tier": {
				"properties": {
					"identifier": {
						"type": "string"
					},
					"isCustom": {
						"type": "boolean"
					},
					"label": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/Type"
					}
				},
				"required": [
					"identifier",
					"isCustom",
					"label",
					"type"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ComparisonGroupIdentifier": {
				"properties": {
					"identifier": {
						"type": "string"
					},
					"isCustom": {
						"type": "boolean"
					},
					"label": {
						"type": "string"
					},
					"members": {
						"items": {
							"$ref": "#/components/schemas/Member"
						},
						"type": "array"
					},
					"owners": {
						"items": {
							"$ref": "#/components/schemas/Member"
						},
						"type": "array"
					},
					"isMain": {
						"type": "boolean"
					},
					"isHidden": {
						"type": "boolean"
					},
					"bottomTier": {
						"$ref": "#/components/schemas/Tier"
					},
					"topTier": {
						"$ref": "#/components/schemas/Tier"
					},
					"type": {
						"$ref": "#/components/schemas/Type"
					}
				},
				"required": [
					"identifier",
					"isCustom",
					"label",
					"members",
					"owners",
					"isMain",
					"isHidden",
					"bottomTier",
					"topTier",
					"type"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IComparisonGroupsIdentifierGetResponse": {
				"properties": {
					"comparisonGroup": {
						"$ref": "#/components/schemas/ComparisonGroupIdentifier"
					}
				},
				"required": [
					"comparisonGroup"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IComparisonGroupsIdentifierBoundsGetResponse": {
				"properties": {
					"boundingBox": {
						"$ref": "#/components/schemas/BoundingBox"
					}
				},
				"required": [
					"boundingBox"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"apiKeyAuth": {
				"type": "apiKey",
				"in": "query",
				"name": "ApplicationKey"
			}
		}
	},
	"info": {
		"title": "Local Government Association API",
		"version": "v1",
		"description": "Access to 1 billion+ metric values for English areas drawn from approximately 50 organisations publishing metrics that describe local authorities and their component areas. (DkxYeiqbmsGVMHOkfbbUGUEomQYkfaiiEJCBDEtC)",
		"contact": {
			"name": "Oraclo LTD",
			"email": "admin@orcalo.co.uk",
			"url": "https://orcalo.co.uk/"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/v1/comparisonGroups": {
			"get": {
				"operationId": "GetComparisonGroup",
				"responses": {
					"200": {
						"description": "IComparisonGroupsGetResponse",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IComparisonGroupsGetResponse"
								}
							}
						}
					}
				},
				"summary": "Comparison groups are area groupings of different types, such as all district local authorities, or all fire organisations in England. Without parameters, all groups, excluding custom groups, are returned",
				"tags": [
					"comparisonGroups"
				],
				"security": [
					{
						"apiKeyAuth": []
					}
				],
				"parameters": [
					{
						"description": "Only the main comparison groups, so includes groupings by authority type, but excludes things like groupings by near neighbours",
						"in": "query",
						"name": "mainOnly",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Groups of which the area is a member, but not necessarily an owner. If area 1 is a near neighbour of area 2, area 1 would be a member of the near neighbour group (along with the other near neighbours) and area 2 would be the owner of the group",
						"in": "query",
						"name": "memberArea",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Group which the area owns, but is not necessairly a member. An area is generally the owner of a group if the group is a main group, such as all local authorities, or is the area to which a group relates, such as near neighbours of an area",
						"in": "query",
						"name": "ownerArea",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Comparison groups of a specified comparison group type for a specified area. For example, which Experian Mosaic near neighbours a particular area has",
						"in": "query",
						"name": "type",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/comparisonGroups/{identifier}": {
			"get": {
				"operationId": "GetComparisonGroupIdentifier",
				"responses": {
					"200": {
						"description": "IComparisonGroupsIdentifierGetResponse",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IComparisonGroupsIdentifierGetResponse"
								}
							}
						}
					}
				},
				"summary": "Returns the details for a particular comparison group",
				"tags": [
					"comparisonGroups"
				],
				"security": [
					{
						"apiKeyAuth": []
					}
				],
				"parameters": [
					{
						"description": "(AllLAInRegion_London)",
						"in": "path",
						"name": "identifier",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/comparisonGroups/{identifier}/bounds": {
			"get": {
				"operationId": "GetComparisonGroupIdentifierBounds",
				"responses": {
					"200": {
						"description": "IComparisonGroupsIdentifierBoundsGetResponse",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IComparisonGroupsIdentifierBoundsGetResponse"
								}
							}
						}
					}
				},
				"summary": "Co-ordinates for the minimum and maxim longitude and latitudes for a particular comparison group",
				"tags": [
					"comparisonGroups"
				],
				"security": [
					{
						"apiKeyAuth": []
					}
				],
				"parameters": [
					{
						"description": "(Core_English_Cities)",
						"in": "path",
						"name": "identifier",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	],
...
}